requirements.txt
absl-py==0.12.0
albumentations==1.0.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
ensemble-boxes==1.0.6
entrypoints==0.3
future==0.18.2
google-auth==1.30.1
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.4.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
iterative-stratification==0.1.6
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mean-average-precision==2021.4.26.0
mistune==0.8.4
mkl-fft==1.3.0
mkl-random
mkl-service==2.3.0
munch==2.5.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numba==0.53.1
numpy
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.2.52
opencv-python-headless==4.5.2.52
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow
pretrainedmodels==0.7.4
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.17.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydicom==2.1.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
Shapely==1.7.1
six
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2021.4.8
timm==0.4.5
torch==1.7.0
torchaudio
torchvision==0.8.1
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
typing-extensions
urllib3==1.26.5
vision-transformer-pytorch==1.0.3
warmup-scheduler==0.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
zipp==3.4.1
omegaconf==2.0.6
pycocotools
imagehash
lxml==4.6.3